fix: make keepalive configuration linux only
Bump deps

+++ pymisp/api.py
@@ -29,15 +29,16 @@ from .mispevent import MISPEvent, MISPAttribute, MISPSighting, MISPLog, MISPObje
 from .abstract import pymisp_json_default, MISPTag, AbstractMISP, describe_types
 
 
-# Enable TCP keepalive by default on every requests
-import socket
-from urllib3.connection import HTTPConnection
-HTTPConnection.default_socket_options = HTTPConnection.default_socket_options + [
-    (socket.SOL_SOCKET, socket.SO_KEEPALIVE, 1),  # enable keepalive
-    (socket.SOL_TCP, socket.TCP_KEEPIDLE, 30),  # Start pinging after 30s of idle time
-    (socket.SOL_TCP, socket.TCP_KEEPINTVL, 10),  # ping every 10s
-    (socket.SOL_TCP, socket.TCP_KEEPCNT, 6)  # kill the connection if 6 ping fail  (60s total)
-]
+if sys.platform == 'linux':
+    # Enable TCP keepalive by default on every requests
+    import socket
+    from urllib3.connection import HTTPConnection
+    HTTPConnection.default_socket_options = HTTPConnection.default_socket_options + [
+        (socket.SOL_SOCKET, socket.SO_KEEPALIVE, 1),  # enable keepalive
+        (socket.SOL_TCP, socket.TCP_KEEPIDLE, 30),  # Start pinging after 30s of idle time
+        (socket.SOL_TCP, socket.TCP_KEEPINTVL, 10),  # ping every 10s
+        (socket.SOL_TCP, socket.TCP_KEEPCNT, 6)  # kill the connection if 6 ping fail  (60s total)
+    ]
 
 try:
     # cached_property exists since Python 3.8
